requirements.txt
dj-database-url==0.5.0
Django==2.1.3
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.3
whitenoise==4.1